requirements.txt
pandas
pyarrow
duckdb
geopandas
shapely
pyyaml
